--- TestConsole/src/main/java/FullWriteAndReadCompare.java
@@ -1,6 +1,0 @@
-public class FullWriteAndReadCompare implements Command{
-    @Override
-    public void execute(String commandStr) {
-
-    }
-}
