--- TestConsole/src/main/java/WriteReadAging.java
@@ -1,6 +1,0 @@
-public class WriteReadAging implements Command{
-    @Override
-    public void execute(String commandStr) {
-        throw new RuntimeException("아직 구현 안됐어요");
-    }
-}
